Create PortsProtocolsServices.ps1
This script creates a complete listing of Ports, Protocols and Services for Microsoft Windows systems.
This script provides IT/IA with a tabular output of currently available and/or active Ports, Protocols and Services.
The output can then be copied and pasted into any document format.
With the results window open, press Ctrl A, then Ctrl C and then paste into any document format.